Add an utils project for unit tests
- Some unit test projects can reference this util project for shared
  utilities
