+++ src/MinimalCover.Domain/Models/FunctionalDependency.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace MinimalCover.Domain.Models
 {
