+++ tests/MinimalCover.Domain.UnitTests/MinimalCover.Domain.UnitTests.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\MinimalCover.Domain\MinimalCover.Domain.csproj" />
+    <ProjectReference Include="..\MinimalCover.UnitTests.Utils\MinimalCover.UnitTests.Utils.csproj" />
   </ItemGroup>
 
 </Project>
