+++ tests/MinimalCover.Domain.UnitTests/Models/FunctionalDependencyTest.cs
@@ -1,26 +1,19 @@
 ﻿using Xunit;
-using System.Linq;
-using MinimalCover.Domain.Models;
+using System;
+using MinimalCover.UnitTests.Utils;
 
 namespace MinimalCover.Domain.UnitTests.Models
 {
   public class FunctionalDependencyTests
   {
-    private static FunctionalDependency ConstructFuncDepFromString(string left, string right, string sep = ",")
-    {
-      var leftAttributes = left.Split(",").Select(a => a.Trim()).ToHashSet();
-      var rightAttributes = right.Split(",").Select(a => a.Trim()).ToHashSet();
-      return new FunctionalDependency(leftAttributes, rightAttributes);
-    }
-
     [Theory]
     [InlineData("a", "c", "a", "c")]
     [InlineData("a,b", "d,e,f", "b,a", "e,f,d")]
     [InlineData("a", "d,c", "a", "c,d")]
     public void DoubleEquals_DifferentFdObjs_ReturnsTrue(string left1, string right1, string left2, string right2)
     {
-      var fd1 = ConstructFuncDepFromString(left1, right1);
-      var fd2 = ConstructFuncDepFromString(left2, right2);
+      var fd1 = FuncDepUtils.ConstructFdFromString(left1, right1);
+      var fd2 = FuncDepUtils.ConstructFdFromString(left2, right2);
       Assert.True(fd1 == fd2, $"{fd1} doesn't equal to {fd2}");
     }
 
@@ -30,7 +23,7 @@ namespace MinimalCover.Domain.UnitTests.Models
     [InlineData("a", "d,c")]
     public void DoubleEquals_SameFdObj_ReturnsTrue(string left, string right)
     {
-      var fd = ConstructFuncDepFromString(left, right);
+      var fd = FuncDepUtils.ConstructFdFromString(left, right);
       var tempFd = fd;
       Assert.True(fd == tempFd, $"{fd} doesn't equal to itself");
     }
@@ -43,8 +36,8 @@ namespace MinimalCover.Domain.UnitTests.Models
     [InlineData("a,b,c", "e,d", "a,x,y", "c,d")]
     public void DoubleEquals_DifferentFdObjs_ReturnsFalse(string left1, string right1, string left2, string right2)
     {
-      var fd1 = ConstructFuncDepFromString(left1, right1);
-      var fd2 = ConstructFuncDepFromString(left2, right2);
+      var fd1 = FuncDepUtils.ConstructFdFromString(left1, right1);
+      var fd2 = FuncDepUtils.ConstructFdFromString(left2, right2);
       Assert.False(fd1 == fd2, $"{fd1} is not supposed to be equal to {fd2}");
     }
 
@@ -56,14 +49,14 @@ namespace MinimalCover.Domain.UnitTests.Models
     [InlineData("a", "c", null)]
     public void DoubleEquals_OtherObj_ReturnsFalse(string left, string right, object obj)
     {
-      var fd = ConstructFuncDepFromString(left, right);
+      var fd = FuncDepUtils.ConstructFdFromString(left, right);
       Assert.False(fd.Equals(obj), $"{fd} is not supposed to be equal to {obj}");
     }
 
     [Fact]
     public void DoubleEquals_OneNullObj_ReturnsFalse()
     {
-      var fd = ConstructFuncDepFromString("a", "b");
+      var fd = FuncDepUtils.ConstructFdFromString("a", "b");
       Assert.False(fd == null, $"{fd} is not supposed to be equal to null");
     }
 
@@ -75,8 +68,8 @@ namespace MinimalCover.Domain.UnitTests.Models
     [InlineData("a,b,c,d,e,f,g", "x,y,z", "a,b,c,d,e,f,g", "x,y,z")]
     public void GetHashCode_CompareTwoHashCodes_ReturnsTrue(string left1, string right1, string left2, string right2)
     {
-      var fd1 = ConstructFuncDepFromString(left1, right1);
-      var fd2 = ConstructFuncDepFromString(left2, right2);
+      var fd1 = FuncDepUtils.ConstructFdFromString(left1, right1);
+      var fd2 = FuncDepUtils.ConstructFdFromString(left2, right2);
       var hc1 = fd1.GetHashCode();
       var hc2 = fd2.GetHashCode();
       Assert.True(hc1 == hc2, $"Hash codes don't match. {hc1} != {hc2}");
