+++ tests/MinimalCover.UnitTests.Utils/MinimalCover.UnitTests.Utils.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\MinimalCover.Domain\MinimalCover.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
